requirements.txt
python-telegram-bot
python-dotenv
requests
quart
fastapi
uvicorn
aiohttp
httpx
telethon
Pillow